Condense mtarget UI, add editor
