+++ libs/gi/db/src/Database/DataManagers/CustomMultiTarget.ts
@@ -27,6 +27,7 @@ export function initCustomTarget(path: string[], multi = 1): CustomTarget {
     path,
     hitMode: 'avgHit',
     bonusStats: {},
+    description: '',
   }
 }
 function validateOptTarget(path: string[]): string[] {
@@ -35,8 +36,15 @@ function validateOptTarget(path: string[]): string[] {
 }
 function validateCustomTarget(ct: unknown): CustomTarget | undefined {
   if (typeof ct !== 'object') return undefined
-  let { weight, path, hitMode, reaction, infusionAura, bonusStats } =
-    ct as CustomTarget
+  let {
+    weight,
+    path,
+    hitMode,
+    reaction,
+    infusionAura,
+    bonusStats,
+    description,
+  } = ct as CustomTarget
 
   if (typeof weight !== 'number' || weight <= 0) weight = 1
 
@@ -63,6 +71,8 @@ function validateCustomTarget(ct: unknown): CustomTarget | undefined {
 
   if (!bonusStats) bonusStats = {}
 
+  if (typeof description !== 'string') description = ''
+
   bonusStats = Object.fromEntries(
     Object.entries(bonusStats).filter(
       ([key, value]) =>
@@ -71,7 +81,15 @@ function validateCustomTarget(ct: unknown): CustomTarget | undefined {
     )
   )
 
-  return { weight, path, hitMode, reaction, infusionAura, bonusStats }
+  return {
+    weight,
+    path,
+    hitMode,
+    reaction,
+    infusionAura,
+    bonusStats,
+    description,
+  }
 }
 export function validateCustomMultiTarget(
   cmt: unknown
