+++ libs/gi/page-team/src/CharacterDisplay/Tabs/TabOptimize/Components/OptimizationTargetSelector.tsx
@@ -1,16 +1,8 @@
 import { useBoolState } from '@genshin-optimizer/common/react-util'
-import { SqBadge } from '@genshin-optimizer/common/ui'
-import { objPathValue } from '@genshin-optimizer/common/util'
-import { useDatabase } from '@genshin-optimizer/gi/db-ui'
-import {
-  DataContext,
-  getDisplayHeader,
-  resolveInfo,
-} from '@genshin-optimizer/gi/ui'
-import type { CalcResult } from '@genshin-optimizer/gi/uidata'
-import { Box, Button, Divider, Stack } from '@mui/material'
-import { useCallback, useContext, useMemo } from 'react'
-import { useTranslation } from 'react-i18next'
+import type { ButtonProps } from '@mui/material'
+import { Button } from '@mui/material'
+import { useCallback } from 'react'
+import OptimizationTargetDisplay from './OptimizationTargetDisplay'
 import type { TargetSelectorModalProps } from './TargetSelectorModal'
 import { TargetSelectorModal } from './TargetSelectorModal'
 
@@ -21,6 +13,7 @@ export default function OptimizationTargetSelector({
   showEmptyTargets = false,
   defaultText,
   targetSelectorModalProps = {},
+  buttonProps = {},
 }: {
   optimizationTarget?: string[]
   setTarget: (target: string[]) => void
@@ -28,8 +21,8 @@ export default function OptimizationTargetSelector({
   showEmptyTargets?: boolean
   defaultText?: string
   targetSelectorModalProps?: Partial<TargetSelectorModalProps>
+  buttonProps?: ButtonProps
 }) {
-  const { t } = useTranslation('page_character_optimize')
   const [show, onShow, onClose] = useBoolState(false)
 
   const setTargetHandler = useCallback(
@@ -39,38 +32,6 @@ export default function OptimizationTargetSelector({
     },
     [onClose, setTarget]
   )
-  const { data } = useContext(DataContext)
-  const database = useDatabase()
-  const displayHeader = useMemo(
-    () =>
-      optimizationTarget &&
-      getDisplayHeader(data, optimizationTarget[0], database),
-    [data, optimizationTarget, database]
-  )
-
-  if (!defaultText) defaultText = t('targetSelector.selectOptTarget')
-
-  const { title, icon, action } = displayHeader ?? {}
-  const node: CalcResult | undefined =
-    optimizationTarget &&
-    (objPathValue(data.getDisplay(), optimizationTarget) as any)
-
-  const invalidTarget =
-    !optimizationTarget ||
-    !displayHeader ||
-    !node ||
-    // Make sure the opt target is valid, if we are not in multi-target
-    (!showEmptyTargets && node.isEmpty)
-
-  const {
-    name,
-    textSuffix,
-    icon: nodeIcon,
-    variant = invalidTarget ? 'secondary' : undefined,
-  } = (node?.info && resolveInfo(node?.info)) ?? {}
-
-  const suffixDisplay = textSuffix && <span> {textSuffix}</span>
-  const iconDisplay = icon ? icon : nodeIcon
   return (
     <>
       <Button
@@ -78,35 +39,13 @@ export default function OptimizationTargetSelector({
         onClick={onShow}
         disabled={disabled}
         sx={{ flexGrow: 1 }}
+        {...buttonProps}
       >
-        {invalidTarget ? (
-          <strong>{defaultText}</strong>
-        ) : (
-          <Stack
-            direction="row"
-            divider={<Divider orientation="vertical" flexItem />}
-            spacing={1}
-          >
-            <Box sx={{ display: 'flex', gap: 1, alignItems: 'center' }}>
-              {iconDisplay}
-              <span>{title}</span>
-              {!!action && (
-                <SqBadge color="success" sx={{ whiteSpace: 'normal' }}>
-                  {action}
-                </SqBadge>
-              )}
-            </Box>
-            <Box sx={{ display: 'flex', alignItems: 'center' }}>
-              <SqBadge
-                color={variant === 'invalid' ? undefined : variant}
-                sx={{ whiteSpace: 'normal' }}
-              >
-                <strong>{name}</strong>
-                {suffixDisplay}
-              </SqBadge>
-            </Box>
-          </Stack>
-        )}
+        <OptimizationTargetDisplay
+          optimizationTarget={optimizationTarget}
+          showEmptyTargets={showEmptyTargets}
+          defaultText={defaultText}
+        />
       </Button>
       <TargetSelectorModal
         show={show}
